+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Modified
 
 - local SSL config error causes ONE to fall back to local mode (as with other connection errors)
+- removed ref2dj method
+- one.util.ses2records now returns empty pandas DataFrame instead of None
 
 ## [1.16.3]